// src/api/models/reservationDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 */
import type { ReservationStatus } from './reservationStatus'

export interface ReservationDto {
  id?: string
  start: string
  end: string
  numberOfPersons: number
  price?: number
  /** @minLength 1 */
  phone: string
  /** @minLength 1 */
  email: string
  /** @minLength 1 */
  firstName: string
  /** @minLength 1 */
  lastName: string
  /** @nullable */
  description?: string | null
  status?: ReservationStatus
}