// src/api/models/reservationStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 */

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReservationStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2
} as const